Add files via upload
A quiz application is an interactive and engaging app that allows users to participate in quizzes, answer questions, and receive feedback on their performance. Quiz apps can be used for a variety of purposes, including private use, business use, and education use. Quizzes can be in a variety of types, including multiple choice and drop-down.